+++ src/components/common/TextEditor.tsx
@@ -1,6 +1,5 @@
 import "react-quill/dist/quill.snow.css";
 import dynamic from "next/dynamic";
-import DOMPurify from "dompurify";
 
 export const createMarkup = (html: string) => {
   return { __html: html };
