+++ src/components/contest/UpdateContest.tsx
@@ -6,6 +6,8 @@ import {
 import { useRouter } from "next/router";
 import Contest from "<@>/types/contest";
 import ProgressIndicator from "../common/ProgressIndicator";
+import IsContestLoading from "./IsContestLoading";
+import IsContestFormLoading from "./isContestFormLoading";
 
 const initialState = {
   id: "",
@@ -120,35 +122,12 @@ const EditContestForm: React.FC = () => {
       }));
       return;
     }
-
-    try {
-      await updateContest(updatedContest).unwrap();
-      // Contest update successful, redirect to contest details
-      router.push(`/admin/contests`);
-    } catch (error) {
-      alert(`An error occurred while updating the contest:, ${error}`);
-    }
+    await updateContest(updatedContest).unwrap();
+    // Contest update successful, redirect to contest details
+    router.push(`/admin/contests`);
   };
   if (isLoading) {
-    return (
-      <div className="rounded-md p-4 w-3/4 mt-28 mx-auto">
-        <div className="animate-pulse flex">
-          <div className="flex-1 space-y-6 py-1">
-            <div className="space-y-3">
-              <div className="pl-12 pr-12">
-                <div className="h-6 bg-slate-100 rounded mt-8"></div>
-                <div className="h-16 bg-slate-100 rounded mt-8"></div>
-                <div className="grid grid-cols-1 sm:grid-cols-2 gap-6">
-                  <div className="h-6 bg-slate-100 rounded mt-8 col-span-1"></div>
-                  <div className="h-6 bg-slate-100 rounded mt-8 col-span-1"></div>
-                </div>
-                <div className="h-6 bg-slate-100 rounded mt-8"></div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    );
+    return <IsContestFormLoading />;
   }
 
   return (
@@ -277,4 +256,3 @@ const EditContestForm: React.FC = () => {
 };
 
 export default EditContestForm;
-
