+++ src/components/home/GetStarted.tsx
@@ -12,7 +12,7 @@ const GetStarted: React.FC = () => {
     (state: RootState) => state.auth.isAuthenticated
   );
   return (
-    <div className="mt-16 h-full mb-16 lg:h-52 my-auto mx-auto bg-primary w-full text-white font-poppins tracking-wider leading-10">
+    <div className="my-16 h-full lg:h-52 my-auto mx-auto bg-primary w-full text-white font-poppins tracking-wider leading-10">
       <h1 className="text-xl sm:text-2xl lg:text-3xl font-bold ml-8 md:ml-28 mt-8">
         Ready To Grow?
       </h1>
@@ -23,7 +23,7 @@ const GetStarted: React.FC = () => {
           your true potential and create a brighter future through technology.
         </p>
         {!isAuthenticated && (
-          <div className="col-span-2 text-xl mt-8 ml-8 md:ml-20 lg:ml-8 py-auto px-auto  mb-8">
+          <div className="col-span-2 text-xl mt-8 ml-8 md:ml-20 lg:ml-8 py-auto px-auto mb-8">
             <button
               className="bg-white text-primary py-3 px-8 ml-2 md:ml-8 rounded-lg hover:scale-110 transition duration-300"
               onClick={handleGetStarted}
@@ -38,4 +38,3 @@ const GetStarted: React.FC = () => {
 };
 
 export default GetStarted;
-
