+++ src/components/my-journey/CheckBox.tsx
@@ -2,22 +2,22 @@ import React from "react";
 import { BiCheck } from "react-icons/bi";
 interface CheckBoxProps {
   isCompleted: boolean;
-  stepper?: boolean;
-  active?: boolean;
+  isStepper?: boolean;
+  isActive?: boolean;
 }
 const CheckBox: React.FC<CheckBoxProps> = ({
   isCompleted,
-  stepper,
-  active,
+  isStepper,
+  isActive,
 }) => {
   return (
     <div
       className={`${
-        stepper ? "w-4 h-4" : "w-5 h-5"
+        isStepper ? "w-4 h-4" : "w-5 h-5"
       } flex shrink-0 rounded-full items-center justify-center  ${
         isCompleted
           ? "bg-primary"
-          : active
+          : isActive
           ? "border-4 border-primary"
           : "border-4"
       }`}
