+++ src/components/my-journey/IsStepLoading.tsx
@@ -1,0 +1,94 @@
+import React from 'react'
+
+const IsStepLoading = () => {
+  return (
+    <div className="animate-pulse w-full">
+            <div className="flex flex-col p-2 w-full gap-8">
+              <div className="grid grid-cols-4 justify-items-start w-full max-w-[68rem] gap-2  text-xs md:text-sm lg:text-base">
+                <div className="flex flex-col w-full gap-2">
+                  <div className="w-full h-3 bg-slate-200 rounded-md"></div>
+                  <div className="flex gap-1 items-center">
+                    <div className="w-4 h-4 rounded-full bg-slate-200"></div>
+                    <div className="w-[60%] h-3 bg-slate-200 rounded-md"></div>
+                  </div>
+                </div>
+
+                <div className="flex flex-col w-full gap-2">
+                  <div className="w-full h-3 bg-slate-200 rounded-md"></div>
+                  <div className="flex gap-1 items-center">
+                    <div className="w-4 h-4 rounded-full bg-slate-200"></div>
+                    <div className="w-[60%] h-3 bg-slate-200 rounded-md"></div>
+                  </div>
+                </div>
+
+                <div className="flex flex-col w-full gap-2">
+                  <div className="w-full h-3 bg-slate-200 rounded-md"></div>
+                  <div className="flex gap-1 items-center">
+                    <div className="w-4 h-4 rounded-full bg-slate-200"></div>
+                    <div className="w-[60%] h-3 bg-slate-200 rounded-md"></div>
+                  </div>
+                </div>
+
+                <div className="flex flex-col w-full gap-2">
+                  <div className="w-full h-3 bg-slate-200 rounded-md"></div>
+                  <div className="flex gap-1 items-center">
+                    <div className="w-4 h-4 rounded-full bg-slate-200"></div>
+                    <div className="w-[60%] h-3 bg-slate-200 rounded-md"></div>
+                  </div>
+                </div>
+              </div>
+
+              <div className="flex flex-col p-4 mt-4">
+                <div className="max-w-sm h-3 bg-slate-200 rounded-sm"></div>
+
+                <div className="flex flex-col p-4 gap-4">
+                  <div className="flex p-2 rounded-md max-w-[44rem] h-fit">
+                    <div className="flex items-start m-4">
+                      <div className="w-6 h-6 rounded-full bg-slate-200"></div>
+                    </div>
+
+                    <div className="flex flex-col gap-3 p-2 w-full">
+                      <div className="w-[30%] h-3 bg-slate-200 rounded-sm"></div>
+                      <div className="flex flex-col gap-1">
+                        <div className="w-full h-3 bg-slate-200 rounded-sm"></div>
+                        <div className="w-[90%] h-3 bg-slate-200 rounded-sm"></div>
+                        <div className="w-full h-3 bg-slate-200 rounded-sm"></div>
+                      </div>
+                    </div>
+                  </div>
+                  <div className="flex p-2 rounded-md max-w-[44rem] h-fit">
+                    <div className="flex items-start m-4">
+                      <div className="w-6 h-6 rounded-full bg-slate-200"></div>
+                    </div>
+
+                    <div className="flex flex-col gap-3 p-2 w-full">
+                      <div className="w-[30%] h-3 bg-slate-200 rounded-sm"></div>
+                      <div className="flex flex-col gap-1">
+                        <div className="w-full h-3 bg-slate-200 rounded-sm"></div>
+                        <div className="w-[90%] h-3 bg-slate-200 rounded-sm"></div>
+                        <div className="w-full h-3 bg-slate-200 rounded-sm"></div>
+                      </div>
+                    </div>
+                  </div>
+                  <div className="flex p-2 rounded-md max-w-[44rem] h-fit">
+                    <div className="flex items-start m-4">
+                      <div className="w-6 h-6 rounded-full bg-slate-200"></div>
+                    </div>
+
+                    <div className="flex flex-col gap-3 p-2 w-full">
+                      <div className="w-[30%] h-3 bg-slate-200 rounded-sm"></div>
+                      <div className="flex flex-col gap-1">
+                        <div className="w-full h-3 bg-slate-200 rounded-sm"></div>
+                        <div className="w-[90%] h-3 bg-slate-200 rounded-sm"></div>
+                        <div className="w-full h-3 bg-slate-200 rounded-sm"></div>
+                      </div>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+  )
+}
+
+export default IsStepLoading
