+++ src/components/my-journey/IsUpcomingContestLoading.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const IsUpcomingContestLoading = () => {
+  return (
+    <div className="animate-pulse w-full flex flex-col p-4 gap-2">
+      <div className="w-full rounded-md  h-20 bg-slate-200"></div>
+      <div className="w-full rounded-md  h-20 bg-slate-200"></div>
+      <div className="w-full rounded-md  h-20 bg-slate-200"></div>
+    </div>
+  );
+};
+
+export default IsUpcomingContestLoading;
